last of the newflow feature helpers
