--- spec/features/profile_screen_spec.rb
@@ -1,37 +1,0 @@
-require 'rails_helper'
-
-feature 'profile screen', js: true do
- describe 'resetting your password' do
-
- end
-
- describe 'adding a password (when you only have a social login ATM' do
-
- end
-
- describe 'adding an email address' do
-
- end
-
-
-  describe 'adding a social login' do
-
-  end
-
-  describe 'deleting your password' do
-
-  end
-
-  describe 'editing your email' do
-
-  end
-
-  describe 'editing your name' do
-
-  end
-
-  describe 'log out' do
-
-  end
-
-end
