+++ spec/features/user_updates_password_spec.rb
@@ -26,7 +26,7 @@ feature 'User updates password on profile screen', js: true do
     find('[data-provider=identity] .add').click
 
     screenshot!
-    newflow_complete_add_password_screen
+    complete_add_password_screen
     screenshot!
     expect(page).to have_no_missing_translations
     expect(page).to have_content(t(:"login_signup_form.how_you_log_in"))
@@ -38,7 +38,7 @@ feature 'User updates password on profile screen', js: true do
 
   scenario "changes existing" do
     find('[data-provider=identity] .edit').click
-    newflow_complete_add_password_screen
+    complete_add_password_screen
     expect(page).to have_no_missing_translations
     expect(page).to have_content(
       ActionView::Base.full_sanitizer.sanitize t(:"users.edit.how_you_sign_in_html")
