+++ spec/handlers/oauth_callback_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe OauthCallback, type: :handler do
     }
 
     let(:user) {
-      create_newflow_user(email)
+      create_user(email)
     }
 
     let(:auth) {
@@ -31,7 +31,7 @@ RSpec.describe OauthCallback, type: :handler do
 
   context 'when no authentication found, but verified user found' do
     before do
-      create_newflow_user(email)
+      create_user(email)
     end
 
     let(:oauth_user_info) {
@@ -94,7 +94,7 @@ RSpec.describe OauthCallback, type: :handler do
 
   context 'when no authentication found and the response from social provider does not include email' do
     before do
-      create_newflow_user(email)
+      create_user(email)
     end
 
     let(:oauth_user_info) {
